refactor(frontend): refactored pages names and extensions
Now, to identify one page how page, must add the extension .page.tsx on components
