+++ frontend/src/components/BaseModal/styles.ts
