+++ frontend/src/components/LayoutBox/index.tsx
