+++ frontend/src/components/LayoutBox/styles.ts
