+++ frontend/src/components/Table/styles.ts
