+++ frontend/src/layouts/BaseLayout.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@/components/shared/Header';
+import { Header } from '@/components/Header';
 import { SessionLayout } from '@/layouts/SessionLayout';
 
 import {
