+++ frontend/src/pages/add/cards/index.page.ts
@@ -1,0 +1,1 @@
+export { default } from './CreateCardsPage'
