+++ frontend/src/pages/add/expenses/index.page.tsx
@@ -1,0 +1,1 @@
+export { default } from './CreateExpensesPage'
