+++ frontend/src/pages/home/components/ExpensesTable/styles.ts
