+++ frontend/src/pages/home/components/Summary/styles.ts
