+++ frontend/src/pages/home/components/SummaryCard/index.tsx
