+++ frontend/src/pages/home/components/SummaryCard/styles.ts
