+++ frontend/src/pages/home/styles.ts
