+++ frontend/src/pages/index.page.ts
@@ -1,0 +1,2 @@
+import { HomePage } from './home';
+export default HomePage;
