+++ frontend/src/pages/login/index.page.tsx
@@ -1,0 +1,1 @@
+export { default } from './LoginPage'
