--- frontend/src/styles/pages/add/cards.style.ts
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const RegisterCardsForm = styled.form`
-	display: flex;
-	flex-direction: column;
-	gap: 1rem;
-	padding: 0 2rem 0 2rem;
-`;
